+++ tests/test_property.rs
@@ -31,14 +31,14 @@ fn property_test() {
 
     {
         let db = DB::open_default(&n).unwrap();
-        let prop_name: std::ffi::CString = properties::STATS.to_owned();
+        let prop_name: properties::PropertyName = properties::STATS.to_owned();
         let value = db.property_value(&prop_name).unwrap().unwrap();
         assert!(value.contains("Stats"));
     }
 
     {
         let db = DB::open_default(&n).unwrap();
-        let prop_name: String = properties::STATS.to_owned().into_string().unwrap();
+        let prop_name: String = properties::STATS.to_owned().into_string();
         let value = db.property_value(&prop_name).unwrap().unwrap();
         assert!(value.contains("Stats"));
     }